Organize hook imports and exports
