+++ client/src/components/NavigationBar.tsx
@@ -1,12 +1,12 @@
-import styles from "../styles/NavigationBar.module.css";
-import animoriLogo from "../resources/animori-logo.png";
-import { Link } from "react-router-dom";
-import PrimaryButton from "./PrimaryButton";
-import { useAuth } from "../hooks/useAuth";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 import FacebookIcon from "@mui/icons-material/Facebook";
 import InstagramIcon from "@mui/icons-material/Instagram";
 import { Button } from "@mui/material";
+import { Link } from "react-router-dom";
+import { useAuth } from "../hooks";
+import animoriLogo from "../resources/animori-logo.png";
+import styles from "../styles/NavigationBar.module.css";
+import PrimaryButton from "./PrimaryButton";
 
 const NavigationBar: React.FC<{}> = () => {
 	const { isAuthenticated, userRole, userID, userEmail, logout } = useAuth();
