+++ client/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useAlert } from "./useAlert";
+export { useAuth } from "./useAuth";
