+++ client/src/hooks/useAuth.ts
@@ -1,5 +1,5 @@
-import { useState, useEffect, useCallback } from "react";
 import Cookies from "js-cookie";
+import { useCallback, useEffect, useState } from "react";
 import { LoginResponse } from "../models";
 
 interface AuthData {
