+++ client/src/pages/Login.tsx
@@ -3,8 +3,7 @@ import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import * as api from "../apiControllers/userController";
 import { AlertMessage, NavigationBar, PrimaryButton } from "../components/";
-import { useAlert } from "../hooks/useAlert";
-import { useAuth } from "../hooks/useAuth";
+import { useAlert, useAuth } from "../hooks";
 import { LoginBody, LoginResponse } from "../models/";
 import styles from "../styles/Login.module.css";
