+++ client/src/pages/Register.tsx
@@ -8,8 +8,7 @@ import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom"; // Import Link for navigation
 import * as api from "../apiControllers/userController";
 import { AlertMessage, NavigationBar, PrimaryButton } from "../components/"; // Import your button component
-import { useAlert } from "../hooks/useAlert";
-import { useAuth } from "../hooks/useAuth";
+import { useAlert, useAuth } from "../hooks";
 import { LoginResponse } from "../models";
 import { RegisterBody } from "../models/RegisterBody";
 import styles from "../styles/Login.module.css"; // Import the CSS file
